# internal/http/router/router.go
package router

import (
	"context"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
	"go.uber.org/zap"

	httpware "github.com/Bengo-Hub/httpware"
	authclient "github.com/Bengo-Hub/shared-auth-client"
	"github.com/bengobox/subscription-service/internal/http/handlers"
	"github.com/bengobox/subscription-service/internal/modules/tenant"
)

// New creates and configures a new chi router.
func New(
	log *zap.Logger,
	healthHandler *handlers.HealthHandler,
	planHandler *handlers.PlanHandler,
	subscriptionHandler *handlers.SubscriptionHandler,
	addonHandler *handlers.AddonHandler,
	featureHandler *handlers.FeatureHandler,
	usageHandler *handlers.UsageHandler,
	serviceChargeHandler *handlers.ServiceChargeHandler,
	billingHandler *handlers.BillingHandler,
	platformHandler *handlers.PlatformHandler,
	rbacHandler *handlers.RBACHandler,
	webhookHandler *handlers.WebhookHandler,
	customAddonHandler *handlers.CustomAddonHandler,
	couponHandler *handlers.CouponHandler,
	usageAdminHandler *handlers.UsageAdminHandler,
	couponAdminHandler *handlers.CouponAdminHandler,
	featureCatalogHandler *handlers.FeatureCatalogHandler,
	productActivationHandler *handlers.ProductActivationHandler,
	backupHandler *handlers.BackupHandler,
	backupDestHandler *handlers.BackupDestinationHandler,
	emailLicenseHandler *handlers.EmailLicenseHandler,
	rateLimitHandler *handlers.RateLimitHandler,
	tokenWalletHandler *handlers.TokenWalletHandler,
	apiKey string,
	authMiddleware *authclient.AuthMiddleware,
	allowedOrigins []string,
	tenantSyncer *tenant.Syncer,
) *chi.Mux {
	r := chi.NewRouter()

	// Standard middleware — use httpware for structured zap logging and recovery,
	// consistent with all other platform services.
	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	r.Use(httpware.RequestID)
	r.Use(httpware.Logging(log))
	r.Use(httpware.Recover(log))

	// CORS
	r.Use(cors.Handler(cors.Options{
		AllowedOrigins:   allowedOrigins,
		AllowedMethods:   []string{"GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "Origin", "X-Request-ID", "X-CSRF-Token", "X-API-Key", "X-Tenant-ID", "X-Tenant-Slug"},
		ExposedHeaders:   []string{"Link", "X-RateLimit-Limit", "X-RateLimit-Remaining", "X-RateLimit-Reset", "Retry-After"},
		AllowCredentials: true,
		MaxAge:           300,
	}))

	// Root redirect to Swagger UI docs
	r.Get("/", func(w http.ResponseWriter, r *http.Request) {
		http.Redirect(w, r, "/v1/docs/", http.StatusFound)
	})
	r.Get("/healthz", healthHandler.Health)
	r.Get("/readyz", healthHandler.Health)

	// Swagger docs (outside /api/v1 — no auth required)
	r.Get("/v1/docs/*", handlers.SwaggerUI)
	r.Get("/api/v1/openapi.json", handlers.OpenAPIJSON)

	r.Route("/api/v1", func(r chi.Router) {
		r.Get("/health", healthHandler.Health)
		r.Get("/healthz", healthHandler.Health) // readiness/liveness probe path used by Helm

		// Public routes (no auth required)
		r.Get("/plans", planHandler.ListPlans)
		r.Get("/plans/{id}", planHandler.GetPlan)
		r.Get("/plans/code/{code}", planHandler.GetPlanByCode)
		// Current subscription Terms & Conditions (shown + accepted in the subscribe flow)
		r.Get("/terms", subscriptionHandler.GetTerms)

		// Public API-token capacity-planning calculator — no tenant/auth needed, so a
		// prospective external-API customer can estimate cost before signing up.
		if tokenWalletHandler != nil {
			r.Post("/tokens/estimate", tokenWalletHandler.Estimate)
		}

		// S2S webhook routes (API key auth handled inside handler)
		if webhookHandler != nil {
			r.Post("/webhooks/treasury/payment-status", webhookHandler.HandleTreasuryPayment)
		}

		// Public read-only routes that accept optional auth for tenant context
		if serviceChargeHandler != nil {
			r.Get("/service-charges/plans", serviceChargeHandler.ListServiceChargePlans)
			r.Get("/service-charges/plans/{code}", serviceChargeHandler.GetServiceChargePlan)
		}

		// Authenticated routes
		r.Group(func(r chi.Router) {
			if authMiddleware != nil {
				r.Use(authMiddleware.RequireAuth)
			}

			// Tenant context (optional for bypass)
			r.Use(httpware.TenantV2(httpware.TenantConfig{
				ClaimsExtractor: func(ctx context.Context) (tenantID, tenantSlug string, isPlatformOwner bool, ok bool) {
					claims, found := authclient.ClaimsFromContext(ctx)
					if !found {
						return "", "", false, false
					}
					// Slug-based platform owner check
					isPO := claims.GetTenantSlug() == "codevertex"
					return claims.TenantID, claims.GetTenantSlug(), isPO, true
				},
				URLParamFunc: chi.URLParam,
				Required:     false,
			}))

			// JIT tenant sync: ensure tenant exists in local DB when slug is in context (avoids "tenant not found" after SSO)
			if tenantSyncer != nil {
				r.Use(func(next http.Handler) http.Handler {
					return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
						slug := httpware.GetTenantSlug(r.Context())
						if slug != "" {
							if _, err := tenantSyncer.SyncTenant(r.Context(), slug); err != nil {
								log.Warn("tenant sync failed during request", zap.String("tenant_slug", slug), zap.Error(err))
							}
						}
						next.ServeHTTP(w, r)
					})
				})
			}

			// Tenant-scoped subscription management
			r.Route("/subscription", func(r chi.Router) {
				r.Get("/", subscriptionHandler.Get)
				r.Post("/", subscriptionHandler.Create)
				r.Put("/plan", subscriptionHandler.ChangePlan)
				// Billing period (MONTHLY/SEMI_ANNUAL/ANNUAL); 6+ months waives the setup fee.
				r.Put("/billing-cycle", subscriptionHandler.UpdateBillingCycle)
				r.Post("/initiate", subscriptionHandler.Initiate)
				r.Get("/settings", subscriptionHandler.GetSettings)
				r.Put("/settings", subscriptionHandler.UpdateSettings)
				// Type-A referral: the tenant's own shareable referral code.
				r.Get("/referral-code", subscriptionHandler.GetReferralCode)
				// Opt-in extra usage (overage) master switch + pending-overage view.
				r.Get("/overage", subscriptionHandler.GetOverage)
				r.Post("/overage/enable", subscriptionHandler.EnableOverage)
				r.Post("/overage/disable", subscriptionHandler.DisableOverage)
			})

			// Billing
			if billingHandler != nil {
				r.Get("/billing", billingHandler.GetBilling)
				r.Get("/billing/invoice-preview", billingHandler.InvoicePreview)
				r.Post("/subscription/payment-method/setup", billingHandler.SetupPaymentMethod)
				r.Post("/subscription/payment-method/confirm", billingHandler.ConfirmPaymentMethod)
				// Pay the subscription automatically by M-Pesa standing order (Ratiba).
				r.Get("/subscription/standing-order", billingHandler.GetStandingOrder)
				r.Post("/subscription/standing-order", billingHandler.RegisterStandingOrder)
			}

			// Credit wallet + coupon redemption
			if couponHandler != nil {
				r.Post("/subscription/coupon/redeem", couponHandler.RedeemCoupon)
				r.Get("/billing/credits", couponHandler.GetCreditWallet)
			}

			// S2S subscription lookup by tenant ID — used by auth-api for JWT enrichment.
			// Accessible via platform API key (service-to-service) or platform owner JWT.
			r.Get("/tenants/{tenant_id}/subscription", subscriptionHandler.GetByTenantID)
			// Per-service-tag subscription view — used by auth-ui billing tab.
			r.Get("/tenants/{tenant_id}/subscriptions", subscriptionHandler.GetServiceSubscriptions)

			// S2S effective rate-limit resolution — used by treasury-api's external eTIMS API
			// middleware to enforce per-tenant requests/minute before counting against Redis.
			if rateLimitHandler != nil {
				r.Get("/tenants/{tenant_id}/rate-limit", rateLimitHandler.GetEffectiveRateLimit)
			}

			// Prepaid API token wallet — GetBalance/GetTransactions/Deduct are S2S (called by
			// treasury-api's external eTIMS API middleware, same API-key-or-platform-owner-JWT
			// gate as rate-limit above); InitiateTopUp is tenant-facing self-serve.
			if tokenWalletHandler != nil {
				r.Get("/tenants/{tenant_id}/tokens/balance", tokenWalletHandler.GetBalance)
				r.Get("/tenants/{tenant_id}/tokens/transactions", tokenWalletHandler.GetTransactions)
				r.Post("/tenants/{tenant_id}/tokens/deduct", tokenWalletHandler.Deduct)
				r.Post("/tenants/{tenant_id}/tokens/refund", tokenWalletHandler.Refund)
				r.Post("/tenants/{tenant_id}/tokens/topup", tokenWalletHandler.InitiateTopUp)
			}

			// S2S expiring subscriptions — used by notifications-api for scheduled expiry warnings.
			r.Get("/subscriptions/expiring", subscriptionHandler.ListExpiring)
			// Switch plan by subscription ID — for billing UI plan change flows.
			r.Put("/subscriptions/{id}/switch-plan", subscriptionHandler.SwitchPlan)

			// Add-on management (PlanFeature-based tenant purchasable addons)
			if addonHandler != nil {
				r.Route("/addons", func(r chi.Router) {
					r.Get("/", addonHandler.ListAddons)
					r.Post("/{feature_code}/purchase", addonHandler.PurchaseAddon)
					r.Delete("/{feature_code}", addonHandler.RemoveAddon)
				})
			}

			// Custom addons — platform-admin configured; tenant read-only view
			if customAddonHandler != nil {
				r.Get("/custom-addons", customAddonHandler.ListMyCustomAddons)
			}

			// Email hosting licenses (tenant-scoped) — see
			// .claude/plans/codevertex-email-hosting-service-plan.md Part 3.
			if emailLicenseHandler != nil {
				emailLicenseHandler.RegisterEmailRoutes(r)
			}

			// Feature gate checks — used by all microservices for Trinity Authorization
			if featureHandler != nil {
				r.Route("/features", func(r chi.Router) {
					r.Get("/", featureHandler.GetEntitlements)
					r.Get("/{code}/check", featureHandler.CheckFeature)
					// Platform-wide feature/limit catalog — powers the admin plan builder.
					if featureCatalogHandler != nil {
						r.Get("/catalog", featureCatalogHandler.ListCatalog)
					}
				})
			}

			// Tenant self-service app activation — distinct from the platform-owner-only
			// /admin/tenants/{tenant_id}/products below: these never take a tenant_id path
			// param, the acting tenant is always the caller's own (resolveTenantID), so a
			// tenant admin can only ever activate/deactivate products for their own tenant.
			if productActivationHandler != nil {
				r.Route("/products", func(r chi.Router) {
					r.Get("/", productActivationHandler.ListMyProducts)
					r.Get("/{product_code}/entitlement", productActivationHandler.CheckProductEntitlement)
					r.Post("/{product_code}/activate", productActivationHandler.ActivateMyProduct)
					r.Post("/{product_code}/deactivate", productActivationHandler.DeactivateMyProduct)
				})
			}

			// Usage reporting — called by microservices to track consumption
			if usageHandler != nil {
				r.Route("/usage", func(r chi.Router) {
					r.Post("/report", usageHandler.ReportUsage)
					r.Get("/", usageHandler.GetUsageSummary)
					r.Get("/summary", usageHandler.GetUsageDashboard)
					r.Get("/alerts", usageHandler.GetAlerts)
				})
			}

			// Service charge plans (GET plans moved to public; tenant-specific remains protected)
			if serviceChargeHandler != nil {
				r.Get("/tenants/{tenantID}/service-charges", serviceChargeHandler.GetTenantServiceCharges)
				// S2S: compute the per-transaction commission for a single payment — used by
				// treasury-api to derive service_charge_amount at intent creation.
				r.Get("/tenants/{tenantID}/service-charge/compute", serviceChargeHandler.ComputeTenantServiceCharge)
			}

			// RBAC routes
			if rbacHandler != nil {
				rbacHandler.RegisterRoutes(r)
			}

			// Tenant-scoped backups (this tenant's data only) — admin-gated: platform owner
			// or a tenant admin holding subscriptions.config.manage.
			if backupHandler != nil {
				r.Group(func(r chi.Router) {
					r.Use(func(next http.Handler) http.Handler {
						return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
							ctx := req.Context()
							allowed := httpware.IsPlatformOwner(ctx)
							if !allowed {
								if claims, ok := authclient.ClaimsFromContext(ctx); ok {
									allowed = claims.HasAnyPermission("subscriptions.config.manage", "subscriptions.config.view")
								}
							}
							if !allowed {
								w.Header().Set("Content-Type", "application/json")
								w.WriteHeader(http.StatusForbidden)
								_, _ = w.Write([]byte(`{"error":"forbidden"}`))
								return
							}
							next.ServeHTTP(w, req)
						})
					})
					backupHandler.RegisterRoutes(r)
					// Per-tenant backup-destination override (mirrors backups off the
					// PVC) — same admin/config gate as the tenant backups routes.
					if backupDestHandler != nil {
						backupDestHandler.RegisterRoutes(r)
					}
				})
			}

			// Platform-default backup destination (OneDrive/GDrive/S3/WebDAV/SFTP/SMB)
			// — platform owner only. Mounted at /api/v1/platform/backups/destination.
			if backupDestHandler != nil {
				r.Route("/platform", func(r chi.Router) {
					r.Use(requirePlatformOwner)
					backupDestHandler.RegisterPlatformRoutes(r)
				})
			}

			// Admin routes for plans (platform admin only)
			r.Route("/admin/plans", func(r chi.Router) {
				r.Use(requirePlatformOwner)
				r.Post("/", planHandler.CreatePlan)
				r.Put("/{id}", planHandler.UpdatePlan)
				r.Delete("/{id}", planHandler.DeletePlan)
			})

			// Admin: feature catalog CRUD (platform owner only)
			if featureCatalogHandler != nil {
				r.Route("/admin/feature-catalog", func(r chi.Router) {
					r.Use(requirePlatformOwner)
					r.Post("/", featureCatalogHandler.UpsertCatalogEntry)
					r.Delete("/{id}", featureCatalogHandler.DeleteCatalogEntry)
				})
			}

			// Admin: service charge plan CRUD
			if serviceChargeHandler != nil {
				r.Route("/admin/service-charges", func(r chi.Router) {
					r.Use(requirePlatformOwner)
					r.Post("/", serviceChargeHandler.CreateServiceChargePlan)
					r.Put("/{id}", serviceChargeHandler.UpdateServiceChargePlan)
					r.Delete("/{id}", serviceChargeHandler.DeleteServiceChargePlan)
				})
			}

			// Admin: tenant and subscription management
			if platformHandler != nil {
				r.Route("/admin/tenants", func(r chi.Router) {
					r.Use(requirePlatformOwner)
					r.Get("/", platformHandler.ListTenants)
					r.Post("/{tenant_id}/subscription", platformHandler.AssignPlanToTenant)
					// Platform-admin: grant/revoke a tenant's blanket subscription exemption.
					r.Patch("/{tenant_id}/exemption", platformHandler.SetTenantExemption)
					// Platform-owner-confirmed, irreversible purge of a dormancy-suspended tenant.
					r.Post("/{tenant_id}/purge-confirm", platformHandler.ConfirmDormancyPurge)
					// Multi-use-case: per-product subscription lines (composite entitlements).
					r.Get("/{tenant_id}/products", platformHandler.ListTenantProducts)
					r.Post("/{tenant_id}/products", platformHandler.AssignProductToTenant)
					r.Delete("/{tenant_id}/products/{product_code}", platformHandler.RemoveProductFromTenant)
					// Add-on features: single named feature_definitions grants independent of any
					// plan/product overlay (per-branch pricing, batch markdown pricing, flash sale, ...).
					// See TenantFeatureGrant's schema doc for why this is distinct from /products above.
					r.Get("/{tenant_id}/feature-grants", platformHandler.ListTenantFeatureGrants)
					r.Post("/{tenant_id}/feature-grants", platformHandler.GrantTenantFeature)
					r.Delete("/{tenant_id}/feature-grants/{feature_code}", platformHandler.RevokeTenantFeature)
					// Subscription invoices (generate / resend / view / download PDF)
					r.Post("/{tenant_id}/subscription/generate-invoice", platformHandler.GenerateSubscriptionInvoice)
					r.Post("/{tenant_id}/subscription/invoice/resend", platformHandler.ResendSubscriptionInvoice)
					r.Get("/{tenant_id}/subscription/invoice", platformHandler.GetSubscriptionInvoice)
					r.Get("/{tenant_id}/subscription/invoice/pdf", platformHandler.DownloadSubscriptionInvoicePDF)
					// Support agreements: standard hosting and support plus special support
					r.Get("/{tenant_id}/support-agreements", platformHandler.ListTenantSupportAgreements)
					r.Post("/{tenant_id}/support-agreements", platformHandler.CreateTenantSupportAgreement)
				})

				r.Route("/admin/subscriptions", func(r chi.Router) {
					r.Use(requirePlatformOwner)
					r.Get("/", platformHandler.ListAllSubscriptions)
					// Edit trial_ends_at, current_period_end, status, plan_code — any combination
					r.Put("/{id}", platformHandler.UpdateSubscription)
					// Convenience alias for status-only updates (kept for backward compatibility)
					r.Put("/{id}/status", platformHandler.UpdateSubscriptionStatus)
				})

				// Support charges: platform receivables, per-charge reprice / waive / offline
				// payment / invoice, and agreement updates.
				r.Route("/admin/support-agreements", func(r chi.Router) {
					r.Use(requirePlatformOwner)
					r.Put("/{id}", platformHandler.UpdateSupportAgreement)
				})
				r.Route("/admin/support-fee-cycles", func(r chi.Router) {
					r.Use(requirePlatformOwner)
					r.Get("/", platformHandler.ListSupportFeeCycles)
					r.Put("/{id}", platformHandler.UpdateSupportFeeCycle)
					r.Post("/{id}/invoice", platformHandler.GenerateSupportFeeCycleInvoice)
				})

				// Platform admin: custom addon CRUD per tenant
				if customAddonHandler != nil {
					r.Route("/admin/tenants/{tenant_id}/custom-addons", func(r chi.Router) {
						r.Use(requirePlatformOwner)
						r.Post("/", customAddonHandler.CreateCustomAddon)
						r.Get("/", customAddonHandler.ListCustomAddonsByTenant)
						r.Patch("/{id}", customAddonHandler.UpdateCustomAddon)
						r.Delete("/{id}", customAddonHandler.DeleteCustomAddon)
					})
				}

				// Platform admin: extend trial for a tenant subscription
				r.Post("/admin/tenants/{tenant_id}/subscription/extend-trial", subscriptionHandler.ExtendTrial)

				// Platform admin: gift credits to a tenant
				if couponHandler != nil {
					r.Post("/admin/tenants/{tenant_id}/credits/gift", couponHandler.GiftCredits)
				}

				r.With(requirePlatformOwner).Get("/platform/stats", platformHandler.GetPlatformStats)

				r.Route("/admin/configs", func(r chi.Router) {
					r.Use(requirePlatformOwner)
					r.Get("/", platformHandler.ListServiceConfigs)
					r.Post("/", platformHandler.CreateServiceConfig)
					r.Put("/{id}", platformHandler.UpdateServiceConfig)
					r.Delete("/{id}", platformHandler.DeleteServiceConfig)
				})
			}

			// Platform admin: per-tenant usage view and manual metric override
			if usageAdminHandler != nil {
				r.Get("/admin/usage/tenants/{tenant_id}", usageAdminHandler.GetTenantUsage)
				r.Put("/admin/usage/tenants/{tenant_id}/override", usageAdminHandler.OverrideMetric)
			}

			// Platform admin: coupon CRUD
			if couponAdminHandler != nil {
				r.Route("/admin/coupons", func(r chi.Router) {
					r.Get("/", couponAdminHandler.ListCoupons)
					r.Post("/", couponAdminHandler.CreateCoupon)
					r.Get("/{id}", couponAdminHandler.GetCoupon)
					r.Patch("/{id}", couponAdminHandler.UpdateCoupon)
					r.Delete("/{id}", couponAdminHandler.DeleteCoupon)
				})
			}
		})
	})

	return r
}

// requirePlatformOwner rejects any caller that is not the platform owner.
func requirePlatformOwner(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if !httpware.IsPlatformOwner(r.Context()) {
			http.Error(w, `{"error":"forbidden"}`, http.StatusForbidden)
			return
		}
		next.ServeHTTP(w, r)
	})
}